refactor: use Parla `structuredContent` via MCP schema mode
Switch `parla_vector_search` to schema mode with an `outputSchema` so the MCP client validates and exposes the server's `structuredContent` directly as tool output. Collapses `parseParlaToolOutput` to a single `safeParse`. Co-locate schema, type and parser in `parla-mcp-tools.ts`.
