+++ apps/backend/src/tools/mcp/parla-mcp-tools.ts
@@ -1,5 +1,5 @@
 import { createMCPClient, MCPClient } from "@ai-sdk/mcp";
-import { tool, type Tool } from "ai";
+import { type Tool } from "ai";
 import { z } from "zod";
 import { captureError } from "../../monitoring/capture-error";
 import { config } from "../../config";
@@ -8,6 +8,36 @@ export interface ParlaMCPToolsResult {
 	tools: Record<string, Tool>;
 	cleanup: () => Promise<void>;
 }
+
+export type ParlaChunkData = {
+	id: number;
+	content: string;
+	page: number;
+	url: string;
+	title: string;
+	source_type: string;
+};
+
+export const parlaResponseSchema = z.object({
+	documentMatches: z.array(
+		z.object({
+			registered_document: z.object({
+				source_url: z.string(),
+				source_type: z.string(),
+				metadata: z.record(z.string(), z.unknown()).nullable(),
+			}),
+			processed_document_chunk_matches: z.array(
+				z.object({
+					processed_document_chunk: z.object({
+						id: z.number(),
+						content: z.string(),
+						page: z.number(),
+					}),
+				}),
+			),
+		}),
+	),
+});
 
 export const parlaMCPTools = async (): Promise<ParlaMCPToolsResult | null> => {
 	let parlaHttpClient: MCPClient | undefined;
@@ -19,16 +49,13 @@ export const parlaMCPTools = async (): Promise<ParlaMCPToolsResult | null> => {
 			},
 		});
 
-		const parlaHttpClientToolSet = await parlaHttpClient.tools();
-
-		// Wrap MCP tools with proper Zod validation
-		// The MCP SDK returns tools with JSON Schema, but the AI SDK needs proper Zod schemas
-		const wrappedTools: Record<string, Tool> = {};
-
-		for (const [toolName, mcpTool] of Object.entries(parlaHttpClientToolSet)) {
-			if (toolName === "parla_vector_search") {
-				wrappedTools[toolName] = tool({
-					description: mcpTool.description || "Vector search tool",
+		// Schema mode (vs "automatic") makes the client validate against
+		// outputSchema and expose the server's structuredContent as the tool
+		// output. Caveat: only the listed tools are returned — others (e.g.
+		// health_check) are dropped.
+		const tools = await parlaHttpClient.tools({
+			schemas: {
+				parla_vector_search: {
 					inputSchema: z.object({
 						query: z.string().describe("The search query"),
 						match_threshold: z
@@ -59,22 +86,13 @@ export const parlaMCPTools = async (): Promise<ParlaMCPToolsResult | null> => {
 							.optional()
 							.describe("Maximum documents to return (default 3)"),
 					}),
-					execute: async (params, options) => {
-						if (mcpTool.execute) {
-							return mcpTool.execute(params, options);
-						}
-						throw new Error("MCP tool execute function not found");
-					},
-				});
-			} else {
-				// For other tools, use them as-is
-				wrappedTools[toolName] = mcpTool as Tool;
-			}
-		}
+					outputSchema: parlaResponseSchema,
+				},
+			},
+		});
 
-		// Return tools and cleanup function instead of closing immediately
 		return {
-			tools: wrappedTools,
+			tools,
 			cleanup: async () => await parlaHttpClient?.close(),
 		};
 	} catch (error) {
@@ -93,3 +111,24 @@ export const parlaMCPTools = async (): Promise<ParlaMCPToolsResult | null> => {
 		return null;
 	}
 };
+
+export function parseParlaToolOutput(output: unknown): ParlaChunkData[] {
+	const parsed = parlaResponseSchema.safeParse(output);
+	if (!parsed.success) {
+		captureError(parsed.error);
+		return [];
+	}
+
+	return parsed.data.documentMatches.flatMap((match) =>
+		match.processed_document_chunk_matches.map((chunkMatch) => ({
+			id: chunkMatch.processed_document_chunk.id,
+			content: chunkMatch.processed_document_chunk.content,
+			page: chunkMatch.processed_document_chunk.page,
+			url: match.registered_document.source_url,
+			title:
+				(match.registered_document.metadata?.["title"] as string | undefined) ??
+				match.registered_document.source_url,
+			source_type: match.registered_document.source_type,
+		})),
+	);
+}
